quantum-backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
httpx==0.28.1
pydantic-settings==2.7.1
sqlalchemy==2.0.36
google-generativeai==0.8.3
reportlab==4.2.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
